pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "uniaudio2"
version = "0.1.0"
description = "UniAudio 2.0: A Unified Audio Language Model with Text-Aligned Factorized Audio Tokenization"
readme = "readme.md"
requires-python = ">=3.9"
license = { text = "To be specified" }
authors = [
    { name = "UniAudio 2.0 Team", email = "" }
]
dependencies = [
    "numpy==2.0.2",
    "torch==2.4.1",
    "torchaudio==2.4.1",
    "torchvision==0.19.1",
    "torchtune==0.4.0",
    "torchao==0.9.0",
    "tqdm==4.67.1",
    "traitlets==5.7.1",
    "traittypes==0.2.3",
    "transformers==4.57.0",
    "tokenizers==0.22.1",
    "ipykernel==6.17.1",
    "einops==0.8.1",
    "accelerate==1.12.0",
    "bitsandbytes==0.49.0",
    "vector-quantize-pytorch==1.27.15",
    "modelscope==1.33.0",
    "soundfile",
    "omegaconf>=2.0",
    "typing_extensions>=4.0",
    "PyYAML",
    "litgpt",
    "diffusers>=0.25.0",
    "safetensors>=0.4.0",
    "huggingface_hub",
    "nnAudio",
    "pytorch_lightning>=2.0.0",
    "scipy",
    "datasets",
    "jiwer",
    "zhconv",
    "zhon",
    "editdistance",
    "regex",
    "more-itertools",
    "absl-py",
    "langdetect",
    "immutabledict",
    "nltk",
    "librosa",
    "laion_clap",
]
# Optional: fairseq often needs install from source for compatibility
# Run after: pip install git+https://github.com/facebookresearch/fairseq.git

[project.optional-dependencies]
dev = ["pytest", "black", "ruff"]
fad = ["audioldm_eval"]
kaldi = ["kaldiio"]

[tool.setuptools.packages.find]
where = ["."]
include = ["evaluation", "llm_models", "llm_modules", "llm_utils", "metrics", "tools", "prompts"]
# exclude = ["tools.kaldi"]  # uncomment to omit Kaldi utils from installed package

classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Operating System :: OS Independent",
]
